refactor(objectpool, testing scripts, bullet scripts): perform refactorings related to bullets
I used th object pool to manage the bullet and wrote related test codes.
