--- Box_Smasher/Assets/Scenes/sih_Scene.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 3ac1d208fb952764ea84c1a6cc26614f
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
