--- Box_Smasher/Assets/Scripts/BS/Runtime/Bullets/ctw_Bullet_Collider_Script.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b75a2364e39b8e14793576f6c695e888
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
