--- Box_Smasher/Assets/Scripts/BS/Runtime/Effects/SparkEffect.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4dd8a1b5786671340a279879647a6a14
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
