--- Box_Smasher/Assets/Scripts/BS/Runtime/Effects/SparkParticle.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 58c3905c03a9ebe41b978e578a40aab6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
