--- Box_Smasher/Assets/Scripts/BS/Runtime/Sound/SoundManger.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3022b94ba61415a4d98cefad19080d3e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
