--- Box_Smasher/Assets/Scripts/BS/sih.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 74bec9ef3802c7c418842061b0badd8b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
