--- Box_Smasher/Assets/Scripts/BS/sih/G.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c97fc6fc0fe677d4ba539d8873f9f811
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
