--- Box_Smasher/Assets/Scripts/BS/sih/G/Materials.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2f87b6be894c0cc4a992d9cd31ce4303
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
