--- Box_Smasher/Assets/Scripts/BS/sih/G/Materials/spakle.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: fd52996b882c7f74da541b2c86d740b7
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
