--- Box_Smasher/Assets/Scripts/BS/sih/Prefap.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: fa9c60feadc94a74fa704dab46632454
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
