--- Box_Smasher/Assets/Scripts/BS/sih/Prefap/spakle0.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: cc6667ff82245134a8d36f925995b3f5
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
