--- Box_Smasher/Assets/Scripts/BS/sih/Prefap/spakle1.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: e543b2725cfe4c243a7fd111748e3ef9
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
