--- Box_Smasher/Assets/Scripts/BS/sih/Script.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0754d635dc4f6dd4ca832181dfdcf9ba
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
