--- Box_Smasher/Assets/Scripts/BS/sih/Script/CameraTracing.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 44b98832a7c0d3b42bab13b9c2e6d974
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
