--- Box_Smasher/Assets/Scripts/BS/sih/soundSource.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b4fb0517414810341be8bbe8fd32bcc3
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
