--- Box_Smasher/Assets/Scripts/BS/sih/soundSource/BGM.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 74b667aa004955149b40553bd5c4f517
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
