--- Box_Smasher/Assets/Scripts/BS/sih/soundSource/BGM/LastRemote.mp3.meta
@@ -1,22 +1,0 @@
-fileFormatVersion: 2
-guid: 816bb6df4100e4f44bd67d719e5d1eff
-AudioImporter:
-  externalObjects: {}
-  serializedVersion: 6
-  defaultSettings:
-    loadType: 0
-    sampleRateSetting: 0
-    sampleRateOverride: 44100
-    compressionFormat: 1
-    quality: 1
-    conversionMode: 0
-  platformSettingOverrides: {}
-  forceToMono: 0
-  normalize: 1
-  preloadAudioData: 1
-  loadInBackground: 0
-  ambisonic: 0
-  3D: 1
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
