+++ Box_Smasher/Packages/manifest.json
@@ -11,7 +11,6 @@
     "com.unity.collab-proxy": "1.2.16",
     "com.unity.ide.rider": "1.1.4",
     "com.unity.ide.vscode": "1.2.3",
-    "com.unity.rendering.hybrid": "0.5.2-preview.4",
     "com.unity.test-framework": "1.1.22",
     "com.unity.testtools.codecoverage": "1.0.0",
     "com.unity.textmeshpro": "2.0.1",
