+++ Box_Smasher/Packages/packages-lock.json
@@ -19,16 +19,17 @@
       "url": "https://packages.unity.com"
     },
     "com.unity.2d.common": {
-      "version": "2.0.2",
+      "version": "2.1.0",
       "depth": 1,
       "source": "registry",
       "dependencies": {
-        "com.unity.2d.sprite": "1.0.0"
+        "com.unity.2d.sprite": "1.0.0",
+        "com.unity.modules.uielements": "1.0.0"
       },
       "url": "https://packages.unity.com"
     },
     "com.unity.2d.path": {
-      "version": "2.0.6",
+      "version": "2.1.0",
       "depth": 1,
       "source": "registry",
       "dependencies": {},
@@ -75,15 +76,6 @@
       "source": "builtin",
       "dependencies": {}
     },
-    "com.unity.burst": {
-      "version": "1.3.0-preview.12",
-      "depth": 2,
-      "source": "registry",
-      "dependencies": {
-        "com.unity.mathematics": "1.1.0"
-      },
-      "url": "https://packages.unity.com"
-    },
     "com.unity.cinemachine": {
       "version": "2.6.4",
       "depth": 0,
@@ -96,34 +88,6 @@
       "depth": 0,
       "source": "registry",
       "dependencies": {},
-      "url": "https://packages.unity.com"
-    },
-    "com.unity.collections": {
-      "version": "0.9.0-preview.6",
-      "depth": 2,
-      "source": "registry",
-      "dependencies": {
-        "com.unity.test-framework.performance": "2.0.8-preview",
-        "com.unity.burst": "1.3.0-preview.12"
-      },
-      "url": "https://packages.unity.com"
-    },
-    "com.unity.entities": {
-      "version": "0.11.1-preview.4",
-      "depth": 1,
-      "source": "registry",
-      "dependencies": {
-        "com.unity.burst": "1.3.0-preview.12",
-        "com.unity.collections": "0.9.0-preview.6",
-        "com.unity.properties": "1.2.0-preview",
-        "com.unity.mathematics": "1.1.0",
-        "com.unity.test-framework.performance": "2.0.8-preview",
-        "com.unity.serialization": "1.2.0-preview",
-        "nuget.mono-cecil": "0.1.5-preview",
-        "com.unity.jobs": "0.2.10-preview.12",
-        "com.unity.scriptablebuildpipeline": "1.6.4-preview",
-        "com.unity.platforms": "0.3.0-preview.4"
-      },
       "url": "https://packages.unity.com"
     },
     "com.unity.ext.nunit": {
@@ -149,96 +113,11 @@
       "dependencies": {},
       "url": "https://packages.unity.com"
     },
-    "com.unity.jobs": {
-      "version": "0.2.10-preview.12",
-      "depth": 2,
-      "source": "registry",
-      "dependencies": {
-        "com.unity.collections": "0.9.0-preview.6",
-        "com.unity.mathematics": "1.1.0"
-      },
-      "url": "https://packages.unity.com"
-    },
     "com.unity.mathematics": {
       "version": "1.1.0",
       "depth": 1,
       "source": "registry",
       "dependencies": {},
-      "url": "https://packages.unity.com"
-    },
-    "com.unity.nuget.newtonsoft-json": {
-      "version": "2.0.0-preview",
-      "depth": 3,
-      "source": "registry",
-      "dependencies": {},
-      "url": "https://packages.unity.com"
-    },
-    "com.unity.platforms": {
-      "version": "0.3.0-preview.4",
-      "depth": 2,
-      "source": "registry",
-      "dependencies": {
-        "com.unity.properties": "1.1.1-preview",
-        "com.unity.properties.ui": "1.1.1-preview",
-        "com.unity.searcher": "4.0.9",
-        "com.unity.serialization": "1.1.1-preview"
-      },
-      "url": "https://packages.unity.com"
-    },
-    "com.unity.properties": {
-      "version": "1.2.0-preview",
-      "depth": 2,
-      "source": "registry",
-      "dependencies": {
-        "nuget.mono-cecil": "0.1.5-preview",
-        "com.unity.test-framework.performance": "1.3.3-preview"
-      },
-      "url": "https://packages.unity.com"
-    },
-    "com.unity.properties.ui": {
-      "version": "1.1.1-preview",
-      "depth": 3,
-      "source": "registry",
-      "dependencies": {
-        "com.unity.properties": "1.1.1-preview",
-        "com.unity.serialization": "1.1.1-preview"
-      },
-      "url": "https://packages.unity.com"
-    },
-    "com.unity.rendering.hybrid": {
-      "version": "0.5.2-preview.4",
-      "depth": 0,
-      "source": "registry",
-      "dependencies": {
-        "com.unity.entities": "0.11.1-preview.4"
-      },
-      "url": "https://packages.unity.com"
-    },
-    "com.unity.scriptablebuildpipeline": {
-      "version": "1.6.4-preview",
-      "depth": 2,
-      "source": "registry",
-      "dependencies": {},
-      "url": "https://packages.unity.com"
-    },
-    "com.unity.searcher": {
-      "version": "4.0.9",
-      "depth": 3,
-      "source": "registry",
-      "dependencies": {},
-      "url": "https://packages.unity.com"
-    },
-    "com.unity.serialization": {
-      "version": "1.2.0-preview",
-      "depth": 2,
-      "source": "registry",
-      "dependencies": {
-        "com.unity.collections": "0.7.0-preview.2",
-        "com.unity.burst": "1.3.0-preview.5",
-        "com.unity.jobs": "0.2.7-preview.11",
-        "com.unity.properties": "1.2.0-preview",
-        "com.unity.test-framework.performance": "1.3.3-preview"
-      },
       "url": "https://packages.unity.com"
     },
     "com.unity.settings-manager": {
@@ -256,16 +135,6 @@
         "com.unity.ext.nunit": "1.0.6",
         "com.unity.modules.imgui": "1.0.0",
         "com.unity.modules.jsonserialize": "1.0.0"
-      },
-      "url": "https://packages.unity.com"
-    },
-    "com.unity.test-framework.performance": {
-      "version": "2.0.8-preview",
-      "depth": 2,
-      "source": "registry",
-      "dependencies": {
-        "com.unity.test-framework": "1.1.0",
-        "com.unity.nuget.newtonsoft-json": "2.0.0-preview"
       },
       "url": "https://packages.unity.com"
     },
@@ -300,7 +169,8 @@
       "depth": 0,
       "source": "builtin",
       "dependencies": {
-        "com.unity.modules.ui": "1.0.0"
+        "com.unity.modules.ui": "1.0.0",
+        "com.unity.modules.imgui": "1.0.0"
       }
     },
     "io.tinu.asepriteimporter": {
@@ -311,13 +181,6 @@
         "com.unity.2d.sprite": "1.0.0"
       },
       "hash": "dce15fde717b175d9a549cde681514f970d90c02"
-    },
-    "nuget.mono-cecil": {
-      "version": "0.1.5-preview",
-      "depth": 2,
-      "source": "registry",
-      "dependencies": {},
-      "url": "https://packages.unity.com"
     },
     "com.unity.modules.ai": {
       "version": "1.0.0",
